refactored electron-main.js
